Fixed alignment issues

+++ Cipher Encryption Project/src/FileUtils.cpp
@@ -16,7 +16,7 @@ std::string *FileUtils::readFromFile(std::string &fileloc)
 	}
 	std::string *content = new std::string;
 	file.seekg(0, std::ios::end);
-	content->reserve(file.tellg());
+	content->reserve(static_cast<const unsigned int>(file.tellg()));
 	file.seekg(0, std::ios::beg);
 	content->assign(std::istreambuf_iterator<char>(file),
 					std::istreambuf_iterator<char>());
